+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 _(add items here for easier creation of next log entry)_
 
+- [Breaking] Change public API function from `locationPickerSuggestions` to `openregisterPickerEngine`.
+
 ## 0.1.0 - 2017-04-28
 
 - [Breaking] Change result format to be `{name, path}` objects.